move configs structs to remove dependency on daemon
 - Moved the following config structs to api/types
   - ContainerRmConfig
   - ContainerCommitConfig
